requirements.txt
django
requests
djangorestframework
psycopg2-binary